Connect the heater directly to battery, instead of 6V regulator
We made this bodge to get more power from the heater. It's not that important that the heater power is regulated. Much more important to get as much power as possible.

The heater is still connected through the P-channel MOSFET, so it will be switched off when the system is switched off.

We made this bodge on all the boards by cutting a trace and replacing it with a jumper wire. This commit updates the project to match.