Add workaround for the mux pins shared with SD
On the M0, the SD card is connected to the same SPI pins that we use for controlling the mux. So after we setup the SD card, we lose control of the mux.

With this workaround, we store the data we would write to the SD card in a 16kB buffer (big enough with a little extra) until we're done with the measurement cycle. Then we write it all at once.